Update patch set 2
Patch Set 2:

> Patch Set 2: -Verified
> 
> > Patch Set 1:
> > 
> > I think the option should allow specifying the mux address, register, and value - so that next time there is a board with a different max we still can use iteflash
> 
> I can add a few extra flags to satisfy this. However, not sure how much this could cover future MUX. It might require a different command sequence. Another concern is that it would be possible to get any values for I2C addr and cmd/data and to practice it without checking the validity.
> 
> My opinion is to have this mux configuration hard-coded rather than to have a flexibility.

I don't think extra flags are required, the -m flag could be accepting parameters:

-m[<port>:<register>:<value>]

you can allow the default to be the bip/apel case.

The security of being able to program other i2c ports on this interface is not an issue, as this operation is allowed if and only if CCD is opened and I2C programming is enabled. When the device is in this state the security is not a concern.

Patch-set: 2
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>