Update patch set 1
Patch Set 1:

I am not sure if the underlying problem is only with the LSM6DSM sensor or it is the case with other sensors.

One thing that concerns me is if there is an adverserial test case that gets added in the future where the test expects to return error for BOGUS ODRs, then the test will fail.

Patch-set: 1
CC: Gerrit User 1302633 <1302633@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>