vboot: increase the CPU clokc during security computation
It was observed that the security verification is the bottleneck for the
EC boot time. In this CL, we reduce this bottleneck by accelerating the
CPU clock during the computation and set it back to normal after.

Because not all chips support/implement the CPU acceleration function,
the calling to clock_enable_module function is enabled only when the
flag "CONFIG_ACCEL_CPU_FOR_SECURITY_COMPUTATION" is defined.

BRANCH=none
BUG=none
TEST=pass "make buildall"
TEST =on npcx7_evb, with related CL, we got ~4x improvement of
security verification time.

Change-Id: I679ca10639aa2978b048e1f23285130f6653649b
Signed-off-by: CHLin <CHLIN56@nuvoton.com>