nocturne: Discharge on AC when changing chg ports
When a charger is plugged in and a higher power charger is plugged into
the other port, nocturne will switch to actively charge from the higher
power charger.  However, during this time, the charger IC is still
switching which puts a large load on the input path.  While opening the
FET to the higher power charger, this load can cause an overcurrent
event and some chargers will latch off due to this overcurrent event.

This commit simply stops the charger IC from switching while we are
switching charge ports.

BUG=chromium:926056
BRANCH=firmware-nocturne-10984.B
TEST=Flash nocturne, plug in Suzy-Q, verify that DUT is charging, plug
in Apple 87W charger, verify DUT switches to Apple charge port and is
charging and no overcurrent event is seen.
TEST=Hibernate nocturne. Plug in charger, verify that DUT boots and is
charging.
TEST=Cutoff nocturne. Plug in charger, verify that DUT boots and is
charging.

Change-Id: I1287c273ad6f7ce7ebfedef6224bd45faa66eb4f
Signed-off-by: Aseda Aboagye <aaboagye@google.com>