ec: Update U2F_ATTEST function for new-style key handles.
The existing implementation of this function was intended for
old-style key handles, and needs to be replaced with a new
implementation for new-style key handles that incorporate
user secrets.

For bonus points, the existing implementatation is actually
broken, and performs no verification, so will blindly attest
to anything, which is Not Good.

BUG=b:124237003
BRANCH=none
TEST=test_that firmware_Cr50U2fCommands

Change-Id: I9b4a487707acf81da39e6495adb42e277f2fdb4a
Signed-off-by: Louis Collard <louiscollard@chromium.org>