Update patch set 1
Patch Set 1:

Btw, what prevents us from passing a couple extra params and getting a signature back as a part of U2F_GENERATE? Then we'll know for sure that we signed the actual handle. And all we need to pass in addition to make it possible seems to be just challenge? The rest (appId, user_secret) is already there, so we can just gen G2FA and sign with it..

Patch-set: 1
Reviewer: Gerrit User 1152319 <1152319@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>