thermal: Don't send thermal warning while in G3
There may be certain states when temperature sensors can't be
read. For instance a thermistor circuit could be tied to a power rail
that is only up when the AP is out of G3. In the function
thermal_control when there are no sensors read, it's calling a
function that adds a warning message to the EC log and sets a host
event. If the AP is in G3.

This CL adds a check for chipset state to guard against adding EC log
messages when it would be an expected state.

BUG=b:123297357
BRANCH=none
TEST=Forced AP into G3 with 'aps' and verified that the log message
"can't read any temp sensors!" is no longer being generated.

Change-Id: Id4d73f58e1d3a6d3c24b9888b385cc5e6f803782
Signed-off-by: Scott Collyer <scollyer@google.com>