WIP, servo_updater: make regex more robust
Seen some logs where due to the regex cutting out too soon, the
servo_updater assumes an update might be needed whereas the version on
the v4/micro is actually already up to date.

This change tightens the regex used to query the version by requiring it
to match some sort of newline character after the version string as
well.

BRANCH=None
BUG=chromium:380567
TEST=None yet

Change-Id: I0b3139a120087d49d6f4bd52267d191716d3b541
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>