usb_chg: Add USB_CHG events for data role set and CC open
The pi3usb9201 bc1.2 chip has both client and host mode support. Other
bc1.2 detection chips support client mode and are triggered by VBUS
detection events. To have better control of the pi3usb9201 and to
prevent d+/d- data interruption, it's preferable to trigger bc1.2 mode
on data role updates as the data role setting is what's most relevant
to legacy bc1.2 modes.

This CL adds 3 new USB_CHG events that can be set in the USB PD
protocol layer when the data role is being set to sink or source and
when a detach event occurs.

BUG=b:124612788
BRANCH=none
TEST=Tested both external chargers, adapters, and charge through
hubs. Verfied that bc1.2 detection is triggered as expected when
chargers are connected and verified that when the type C port is in
source mode, then the pi3usb9201 is in host CDP mode.

Change-Id: I9dbb441c6e226a7836220fc6abcf723a46b1a349
Signed-off-by: Scott Collyer <scollyer@google.com>