bc12: pi3usb9201: Add 5V rail control for bc1.2 detection
In cases where the pi3usb9201 is powered by a 5V rail which depends on
the AP power state, bc1.2 detection may not function when the AP is in
G3. This CL adds code to enable the PP5000 rail prior to setting the
pi3usb9201 into either host or client mode.

BUG=b:122265772
BRANCH=none
TEST=Verfied that when pi3usb9201 is accessed while AP is in G3 that
the i2c messages no longer fail. Prior to this change, bc1.2 detection
following EC reboot was always failing.

Change-Id: Ic7113ac8086237e109b4b08bca985127c3bacdf4
Signed-off-by: Scott Collyer <scollyer@google.com>