inline assembly: Use input operands for symbols
When compiling with LTO, the linker doesn't care for the symbols in the
assembly but works with the compiler's internal representation that's
stored in the object files alongside the "legacy" raw code.

While doing this, inline assembler statements are pasted as strings
without parsing for symbols that may be referenced there. This can lead
to the linker discarding objects early on that are later (when
processing the assembler code) found to be used.

The result is confusing "missing symbol" errors.

By moving the symbol names out of the assembler text and into the inline
assembly input operands, they're made available to the compiler and
therefore to its internal representation. That way everything works out.

BUG=b:65441143,b:65253310
BRANCH=cr50
TEST=the branch instructions still point to the right function in the
ec.*.elf disassembly

Change-Id: I1d311eebeb6d1dadd616e832a5c97cc244e84bcb
Signed-off-by: Patrick Georgi <pgeorgi@google.com>