cr50: use board prop to mask straps if config is ambiguous
Cr50 tries to determine SPI vs I2C by checking that there are pullups on
only the I2C or SPI straps. If there are pullups on both, it can't use
the straps to determine SPI vs I2C, so it just defaults to the default
board properties (I2C and PLT_RST_L).

This change uses last board properties to determine I2C vs SPI if the
straps can't be used to make this determination. The actual strap pins
are only connected to the pullup or pulldown being used for the straps.
It's pretty likely these strap readings are correct. The unused straps
are connected to the AP which can interfere with these readings if it
has a pullup on any of these signals. This change uses the last board
property configuration to determine I2C vs SPI and mask unused straps to
try and remove signals the AP is interfering with. If this masked config
isn't valid, then cr50 will still fall back to using I2C and PLT_RST_L.

BUG=b:124777847
BRANCH=cr50
TEST=flash mistral. Reboot cr50. Make sure it doesn't switch to I2C
config

Change-Id: I409e2f3ab1339aafe450b35259adc3c4c5d870ae
Signed-off-by: Mary Ruthven <mruthven@google.com>