PD: Set Rp consistently on initial connection, increasing later
This change adds a call to the auto toggle state machine to ensure
that the Rp cached value is set to CONFIG_USB_PD_PULLUP before
setting Rp on the CC line. Without this, we may present 3.0 A Rp
briefly between the auto toggle state and the set_state call,
causing potential CCD mode detection.

This change also ensures that the correct Rp to represent our current
capabilities will be presented once in Attached.SRC state.  Previously,
this value would be cached but not presented since the CC line wouldn't
be reset.

BRANCH=octopus
BUG=b:123063171, b:124531988
TEST=on orbatrix with an HDMI dongle, added debug prints into the
tcpci set_role_ctrl function to ensure that any Rp values set before
the partner was attached were to 1.5 A and Rp was set to 3.0 A after
the connection was stable, regardless of whether the dongle had been
plugged into port 1 previously

Change-Id: Iba38f9c07d85c1eb03ae1b8b3e476197a9841121
Signed-off-by: Diana Z <dzigterman@chromium.org>