usb_pd: Adding USBC cable attribute detection
When a discover identity command is sent with a SOP prime
packet, the cable plug of an emark cable responds with the
cable attribute.

Source runs the cable detection prior to sending the source
capability messages and the sink selects the cable attributes
offered by the source. Reference-USB_PD_R3_0 Section 4.4

Added a variable cable_type that stores the cable attribute on
connection and resetting it when the cable is disconnected.
Also added a console command that gives the cable type of the
connected cable.

BUG=b:112311321
BRANCH=None
TEST=Verified on Dragonegg, able to get correct USBC cable attribute

Change-Id: Id51e9fcb7299b00f9304932a2e7123992f506050
Signed-off-by: Ayushee <ayushee.shah@intel.com>