usb_pd: Adding USBC cable attribute detection.
When a discover identity command is sent with a SOP prime
packet, the cable plug of an emark cable responds with the
cable attribute.

Added a variable cable_type that stores the cable attribute on
connection and resetting it when the cable is disconnected.
Also added a console command that gives the cable type of the
connected cable.

BUG=None
BRANCH=None
TEST=Verified on Dragonegg, able to get correct USBC cable attribute

Change-Id: Id51e9fcb7299b00f9304932a2e7123992f506050
Signed-off-by: Ayushee <ayushee.shah@intel.com>