Update patch set 1
Patch Set 1: Code-Review-2

(4 comments)

This patch may be good itself but I'm casting -2 to stop it from being merged.

There is a gap between the Rp states held by a TCPC and a EC. Can we synchronize them? That is, when we set it to 1.5A in auto-toggle, we also set vbus_rp[port] to 1.5A. Yes, that means we have to do power distribution on connect (instead of disconnect as we currently do).

Patch-set: 1
Reviewer: Gerrit User 1106949 <1106949@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-2