Update patch set 1
Patch Set 1:

(3 comments)

> Patch Set 1: Code-Review-2
> 
> (4 comments)
> 
> This patch may be good itself but I'm casting -2 to stop it from being merged.
> 
> There is a gap between the Rp states held by a TCPC and a EC. Can we synchronize them? That is, when we set it to 1.5A in auto-toggle, we also set vbus_rp[port] to 1.5A. Yes, that means we have to do power distribution on connect (instead of disconnect as we currently do).

I'm not familiar with vbus_rp since I don't believe we currently use a structure like that in octopus.  Is there a call that can be used to set this on boards which do use it?  

The primary issue I'm interested in solving with this CL is the difference between whether we present 1.5 A or 3.0 A on plug in, which currently varies based on whether you re-plug to the same port, or move a dongle from one port to the other.

There does seem to be a disconnect between how the charge manager sees the Rp values and which values are actually being presented on the CC lines.  In this case, we're able to use that behavior to avoid presenting 3.0 A and possibly triggering CCD.  But in general, it's definitely a behavior that takes a while to wrap your head around.

Patch-set: 1
Reviewer: Gerrit User 1287197 <1287197@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>