kukui: Pulse PMIC_FORCE_RESET_ODL on external reset
If the reset cause is external, it is wise to reset the PMIC
registers, which can be achieved by pulsing the force reset signal
to PMIC.

We found in previous experiments that 5ms is often enough to reset
the PMIC, but that occasionally the PMIC would auto-restart if
the pulse is so short (so we pulse for 10s when we force shutdown
the AP).

Here, we pulse for 100ms, which is an acceptable delay in these
situations, and hopefully long enough to recover from bad PMIC
settings.

BRANCH=none
BUG=b:117747023
TEST=Boot kukui, external reset cause the PMIC to power cycle,
     measured using PMIC_EC_RESETB.
TEST=AP normal reboot does not cause such a reset.
TEST=See that most PMIC registers are reset after force reset
     is used (see bug).

Change-Id: Iddbc20506890b3f904c36642515a0502d90462a3
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>