ec: Refactor get_rtc_diff to return unsigned values
get_rtc_diff in stm32 previously returned a signed value, but the result
could never be negative. This changes the function to return an unsigned
value in a safe manner, and updates its corresponding usage.

--> Work in progress, do not merge <--

I made a test for get_rtc_diff, but I could not figure out a great way
to link in the code from the stm32 chip to use it in the test, and had
to copy/paste the corresponding function from the stm32 test into the
code (obviously, not ideal). I am looking for advice on how to test as a
certain $BOARD, or if making that happen is not very easy, it is
probably best just to remove my test, and call it good.

So anyway, advice appreciated.

BUG=chromium:769407
BRANCH=none
TEST=make buildall -j
TEST=made unit test for get_rtc_diff; test passed (see note above)

Change-Id: Iec1b274294d8f78eaa0ee66d3bd63ce69536e747
Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>