Ampton: Correct HPD GPIOs
Since ampton is using parade as its USB mux on both ports, both HPD
lines are inputs.  This misconfiguration didn't cause any
issues with the EC driving the GPIOs high because they were set as open
drain, but does make it impossible to read the HPD level with "gpioget".

BRANCH=octopus
BUG=None
TEST=on apel, verified that gpioget now returns correct level for HPD
pins when displays are plugged and unplugged

Change-Id: I0ef628cf7ea68884f97a080ba825d780f7feee41
Signed-off-by: Diana Z <dzigterman@chromium.org>