Update patch set 1
Patch Set 1:

As discussed on the bug, I think we need this for mistral. I think it would be good to implement only reading board properties after power-on reset, but it isn't necessary for mistral, so we can implement it later.

Patch-set: 1
Reviewer: Gerrit User 1136059 <1136059@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>