Update patch set 6
Patch Set 6:

(2 comments)

Please split this into 2 CLs, one that modifies the driver and adds the new config option. Then a 2nd that enables this option in baseboard.h.

Also, is the prochot profile exist for the bq25703 as well? If so we should make the same change for that driver.

Patch-set: 6
Reviewer: Gerrit User 1123814 <1123814@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>