usbpd: Remove pd_state_names if not used
Currently, pd_state_names is included in all builds. This change makes
it included only in such builds as:

- CONFIG_USB_PD_DEBUG_LEVEL is defined to 1+ OR
- CONFIG_USB_PD_DEBUG_LEVEL is not defined

The 'pd dump' console command will continue to print & change debug level
in a normal build. If CONFIG_USB_PD_DEBUG_LEVEL is explicitly defined,
the command will only print the current debug level (as expected).

Enum pd_state doesn't have to be different between different
configurations (DRP, no-DRP, etc.). Actually, we should just stick to a
single definition (and append more to the tail to grow) so that the
consumers can translate the numbers to the names.

After the change with CONFIG_USB_PD_DEBUG_LEVEL = 0,
  93520 Mar 11 12:53 build/flapjack/RO/ec.RO.flat
 104836 Mar 11 12:54 build/flapjack/RW/ec.RW.flat

After the change with CONFIG_USB_PD_DEBUG_LEVEL = 2,
  94672 Mar 11 12:54 build/flapjack/RO/ec.RO.flat
 106004 Mar 11 12:55 build/flapjack/RW/ec.RW.flat

Whatever script, test, daemon, etc. is relying on the text
representation of the PD state needs to be fixed as well.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=chromium:937427
BRANCH=none
TEST=buildall. Verify 'pd dump' works like bofore in a normal build.

Change-Id: Id66634240050103253d69f257ee6f717252fc417