mt_scp: Check for IPI_NS_SERVICE in ipi_send.
Since the IPI_NS_SERVICE is send-only, there's no corresponding IPI
handler on the SCP side, and the IPI ID is 0xFF which is outside of the
table range. Check that the IPI ID is not IPI_NS_SERVICE before checking
if the IPI is a wakeup source.

BUG=b:126535633
TEST=build
BRANCH=none

Change-Id: Id8b67be235afd9606b8d09257c7938a2695bcc0b
Signed-off-by: Pi-Hsun Shih <pihsun@chromium.org>