getversion.sh: check directories existence before trying use them
When gitversion.sh is invoked when building gsctools by the ec-utils
ebuid, the tpm2 and cryptc directories are not present in the
compilation space, and in fact those directories' contents do not
matter for gsctools.

Avoid getversion.sh generating error messages by not trying to examine
directories which do not exist.

BRANCH=none
BUG=none

TEST='sudo emerge ec-utils' does not produce getversion.sh errors
     about not being able to change directory to
     ../../third_party/{tpm2,cryptoc}, but the version string is still
     available:

  $ gsctool -v
  Version: 1.1.9999-6e06c78, built on  by @eskimo.mtv.corp.google.com

     note that date is also not included in this case, but this is a
     separate unrelated issue.

Change-Id: Iadd7615f980e9444c539b5e04afca3f23d080174
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>