Update patch set 9
Patch Set 9:

This change is affecting the 'fan_is_stalled()' function when the target RPM is 0.

Actual:    0 rpm
Target:    0 rpm
Duty:   0%
Status: 0 (not spinning)
Mode:   rpm
Auto:   yes
Enable: yes
Power:  yes
> [60.195219 Fan 0 stalled!]
[61.197603 Fan 0 stalled!]
[62.199852 Fan 0 stalled!]
[63.202208 Fan 0 stalled!]
[64.204349 Fan 0 stalled!]
[65.206663 Fan 0 stalled!]

This is on a Hatch system where fan is being controlled by EC. Note if I force target RPM to be non-zero then the Fan 0 stalled messages do go away. So this seems to be related to be case that only shows up with desired RPM is 0 (fan should be off case)

Patch-set: 9
CC: Gerrit User 1123814 <1123814@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>