USB PD: Re-discover port partner when a contract was in place
When the pd_task starts up with an explicit contract previously in
place, re-check the partner's identity.  This will happen automatically
when we EC reset into RO since pd_chipset_startup sets the flag, but for
development systems jumping from RO to RW, it needs to be set as well.

The flag doesn't remedy display on all dongles and monitors, but does
improve behavior.  This should be acceptable, since the scenario won't
come up on customer systems which are locked.

BUG=b:125552060
BRANCH=octopus
TEST=on unlocked octopus board, plugged in powered HDMI dongle and hit
refresh+power and confirmed display worked after RO->RW jump

Change-Id: Idcde6f04deeb8f409a9b4d0a4b3fc924bdb644c7
Signed-off-by: Diana Z <dzigterman@chromium.org>