Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > Ah, yeah that is correct.
> > > > 
> > > > && Yes, I think consolidating it would be good so we don't miss an edge somewhere. The discovery states sound good to me.
> > > 
> > > Thinking about it overnight, my only concern with the discovery states would be that we'd be increasing how often we start our VDM sequence if we set the flag after every soft reset.  Would that be okay?
> > 
> > That should be okay as a SoftReset keeps the physical layer the same, but resets the protocol layer, so we'd have to go down that path again anyways.
> 
> From the PD spec (6.4.4.4.2 "Enter Vendor Mode/ Exit Vendor Mode Process"), it looks like we only exit alternate modes on a hard reset, so currently the soft reset logic doesn't trigger the discovery path again.
> 
> I tried putting the flag set into set_state on entering either SNK or SRC discovery, then plugging in an HDMI dongle and running "pd 0 soft".  Testing with two different monitors and two different dongles, one monitor reliably won't display anything after re-discovery and the other will just black out briefly and come back.  So the current, somewhat disjointed, setting of this flag may be more desirable to work with a variety of peripherals.

Thanks for the pointer. It's a bit surprising to me, but at the same time I can see how it makes sense. Since we leave the physical layer alone, the SoftReset message would be transparent to the DP Alt mode as the pins are still configured for DisplayPort usage. But then, it seems to me that we need to also keep track of what mode we were in prior to the SoftReset. A lot of monitors and dongles don't like it when you EnterMode twice, and that's what may have been happening.

So my next question is why does patchset 1 work then?

Here's another thought: should we exit the mode, and then set the check identity flag in the sysjump case?

Patch-set: 1
Reviewer: Gerrit User 1125984 <1125984@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>