Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > (1 comment)
> 
> Add anx7447_mux_init()in anx7447_int() may not be a proper way since both type-c ports have the type-c cable connected will run anx7447_int. That means anx7447_mux_set() will be run when Parada port is in active mode but Analogix port is empty.

As mentioned in the meeting, in my opinion it is fined because it just reinitializes the mux. If the device is plugged in, the mux value should be reconfigured to the correct value. Please double check with analogix to see if it brokes the internal state machine.

Patch-set: 4