Update patch set 2
Patch Set 2:

I think you should reduce the width a little bit, but otherwise it seems fine to me. From

  [12 Mar 2019 14:05:15.788918] [4.179 HC 0x67 err 9]

I would suggest going to

  2019-03-12 14:05:15 [4.179 HC 0x67 err 9]

There's little reason to have a subsecond timestamp from the host, and it looks somewhat odd to have both parts in square brackets. Switching to YYYY-MM-DD also moves it closer to what syslog uses.

However, I still think your implementation doesn't quite work right, because the timestamp does not match the line it is printed with (it matches the previous line). You can't just replace \n with a timestamp because the \n is output at the end of a log message, not at the beginning. Rather, you should do something like set a newline_seen variable whenever you see a \n, and then whenever it tries to output the next character after newline_seen is true, you insert the current timestamp beforehand and set it to false again.

(I think there may be more weird cases to consider if you also take console input into account. I'm not quite sure how that works, e.g. whether the "> " is printed by ec3po or by the EC itself... but you can get situations where the "> " is already printed, but the user isn't done typing a command, and then the EC sends output in between. It would be neat if we could insert a newline and a new timestamp in that case so that EC output would always be prefixed with the correct timestamp no matter what the user types into the console. But that edge case might also just not be important enough to worry about... I think we mostly need this to work when the console is being passively recorded.)

Patch-set: 2
Reviewer: Gerrit User 1002133 <1002133@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>