kukui: scp: calibrate ULPOSC1&2
ULPOSC generates clock for SCP core and peripherals. The calibration
process adjust 2 values, div and cali. Both values are positive
correlated to OSC frequency. The frequency function is:
  f(div, cali) = k1 * (div + k2) / R(cali) * C
Where:
  R(cali) = k3 / (1 + k4 * (cali - k4))

The actual frequency is not linear to cali parameter. This change
selects the div that generates closest frequency when cali == 32. And
then adjust cali to get better output.

BRANCH=none
BUG=b:120176040,b:120169529
TEST=manual
  check SCP console command:
  > ulposc
  ULPOSC1 frequency: 248 MHz
  ULPOSC2 frequency: 330 MHz

Change-Id: I1a6b9d1ad140cd67b1fdecebf14b61cad88b2c4f
Signed-off-by: Rong Chang <rongchang@chromium.org>
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>