Flapjack: Cut off battery when USB-C board isn't connected
Currently, when a battery is connected, the device boots immediately.
This makes assembly process unreliable.

This patch makes EC cut off a battery at start-up if USBC_THERM reads
about the pull-up voltage because that indicates a USB-C board isn't
connected, which means the devices is being assembled.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b/124540674
BRANCH=none
TEST=Verify a battery is cut off when a USB-C board isn't connected.
Verify a battery isn't cut off when a USB-C board is connected.

Change-Id: Iadc756f8705abe083a5d56b5d8ad080dbff15b6a