cr50: Update to VENDOR_CC_GET_PWR_BTN
Change the VENDOR_CC_GET_PWR_BTN behavior to report the live state of
the power button.  On Wilco devices, the AP needs to poll for both a
power button press and release prior to booting an untrusted OS.

The falling edge of the power button is still latched, so the first call
to VENDOR_CC_GET_PWR_BTN command will return a button state of 1 if this
call is made within 10 seconds.  Once the latched state is clear,
VENDOR_CC_GET_PWR_BTN returns the live state of the power button.

BUG=b:128431787
BRANCH=cr50
TEST=Added test code into vboot recovery screen to poll
VENDOR_CC_GET_PWR_BTN command for a power button press and a power
button release.

Change-Id: I8bc34e4e6309f6b3bd8020655370eb50032a00f1
Signed-off-by: Keith Short <keithshort@chromium.org>