Update patch set 3
Patch Set 3:

I uploaded PatchSet3 for different approach. What are your thoughts?

It was going to be to be a lot of comments to explain, so it was easier just to upload a new patch set for discussion. I didn't not complete the patchset. I am not trying to take over the CL :)

Patch-set: 3
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>